Deploy back

+++ src/routes/reviews/reviews-functions.js
@@ -16,7 +16,7 @@ const crear = async (req, res, next) => {
     };
     let date = `${today.getFullYear()}-${month(today)}-${today.getDate()}`;
 
-    const todayPosts = userDb.reviews.filter((r) => {
+    const todayPosts = userDb.reviews?.filter((r) => {
       return r.createdAt == date;
     });
 
@@ -272,20 +272,29 @@ const deleteReview = async (req, res, next) => {
 const likeReview = async (req, res, next) => {
   try {
     const { userId, reviewId } = req.params;
-    const userDb = await User.findByPk(userId);
-    const hasLike = await userDb.hasLikes(reviewId);
-
-    if (hasLike) {
-      await userDb.removeLikes(reviewId);
-      res.send("Diste dislike a esta review");
+    const userDb = await User.findOne({where: {id:userId}}); // findByPk(userId)
+    console.log("USUARIO", userDb)
+    if (userDb) {
+      try {
+        const hasLike = await userDb.hasLikes(reviewId);
+        if (hasLike) {
+          await userDb.removeLikes(reviewId);
+          res.send("Quitaste tu a esta review");
+        } else {
+          await userDb.addLikes(reviewId);
+          res.send("Diste like a esta review");
+        }
+      } catch (error) {
+        console.log(error)
+      }
     } else {
-      await userDb.addLikes(reviewId);
-      res.send("Diste like a esta review");
+      res.status(404).json("Usuario no encontrado");
     }
   } catch (error) {
-    next(error);
-  }
+    console.log(error);
+  };
 };
+
 module.exports = {
   crear,
   modificar,
